# docs/callback/swagger.yaml
basePath: /callback/v1
definitions:
...
  request.Avatar:
    properties:
      key:
        description: 前端透传给后端用于保存avatar，例如：custom-upload/avatar/abc/def.png
        type: string
      path:
        description: 前端请求地址，例如：/v1/static/avatar/abc/def.png (请求非必填)
        type: string
    type: object
...
  request.SearchAcquiredSkillListReq:
    properties:
      skillIdList:
        items:
          type: string
        type: array
    required:
    - skillIdList
    type: object
  request.SearchBuiltinSkillListReq:
    properties:
      skillIdList:
        items:
          type: string
        type: array
    required:
    - skillIdList
    type: object
  request.SearchCustomSkillListReq:
    properties:
      skillIdList:
        items:
          type: string
        type: array
    required:
    - skillIdList
    type: object
...
  response.CallbackAcquiredSkillDetail:
    properties:
      author:
        type: string
      avatar:
        $ref: '#/definitions/request.Avatar'
      desc:
        type: string
      name:
        type: string
      objectPath:
        type: string
      skillId:
        type: string
    type: object
  response.CallbackAcquiredSkillDetailListResp:
    properties:
      skillList:
        items:
          $ref: '#/definitions/response.CallbackAcquiredSkillDetail'
        type: array
    type: object
...
  response.CustomSkillDetailListResp:
    properties:
      skillList:
        items:
          $ref: '#/definitions/response.CustomSkillListDetail'
        type: array
    type: object
  response.CustomSkillListDetail:
    properties:
      author:
        type: string
      avatar:
        $ref: '#/definitions/request.Avatar'
      desc:
        type: string
      name:
        type: string
      objectPath:
        type: string
      skillId:
        type: string
      variables:
        items:
          $ref: '#/definitions/response.SkillVariable'
        type: array
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  response.SkillDetail:
    properties:
      author:
        type: string
      avatar:
        $ref: '#/definitions/request.Avatar'
      desc:
        type: string
      name:
        type: string
      skillId:
        type: string
      skillMarkdown:
        type: string
      skillPath:
        type: string
      variables:
        items:
          $ref: '#/definitions/response.SkillVariable'
        type: array
    type: object
  response.SkillDetailListResp:
    properties:
      skillList:
        items:
          $ref: '#/definitions/response.SkillDetail'
        type: array
    type: object
  response.SkillVariable:
    properties:
      desc:
        type: string
      id:
        type: string
      name:
        type: string
      variableKey:
        type: string
      variableValue:
        type: string
    type: object
...
info:
  contact: {}
  title: AI Agent Productivity Platform - Callback
  version: v0.0.1
paths:
...
  /callback/v1/skill/acquired/list:
    post:
      consumes:
      - application/json
      description: 获取我添加skill详情列表
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SearchAcquiredSkillListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CallbackAcquiredSkillDetailListResp'
              type: object
      summary: 获取我添加skill详情列表
      tags:
      - skill
  /callback/v1/skill/builtin/list:
    post:
      consumes:
      - application/json
      description: 获取内置skill详情列表
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SearchBuiltinSkillListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.SkillDetailListResp'
              type: object
      summary: 获取内置skill详情列表
      tags:
      - skill
  /callback/v1/skill/custom/list:
    post:
      consumes:
      - application/json
      description: 获取自定义skill详情列表
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SearchCustomSkillListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CustomSkillDetailListResp'
              type: object
      summary: 获取自定义skill详情列表
      tags:
      - skill
...
swagger: "2.0"